fix: FreeBSD build - int64/uint64 type mismatch in statfs
- tmpfs.go: Convert stat.Blocks/Bavail/Bfree to int64 for cross-platform math
- large_db_guard.go: Same fix for disk space calculation
- FreeBSD uses int64 for these fields, Linux uses uint64

+++ internal/restore/large_db_guard.go
@@ -674,7 +674,8 @@ func (g *LargeDBGuard) CheckTmpfsAvailable() *TmpfsRecommendation {
 			continue
 		}
 
-		freeBytes := stat.Bavail * uint64(stat.Bsize)
+		// Use int64 for cross-platform compatibility (FreeBSD uses int64)
+		freeBytes := uint64(int64(stat.Bavail) * int64(stat.Bsize))
 
 		// Skip if less than 512MB free
 		if freeBytes < 512*1024*1024 {
